final rewrites
