+++ trulens_eval/trulens_eval/app.py
@@ -33,16 +33,16 @@ from trulens_eval.schema import Select
 from trulens_eval.tru import Tru
 from trulens_eval.utils.pyschema import Class
 from trulens_eval.utils.serial import all_objects
-from trulens_eval.utils.serial import callable_name
-from trulens_eval.utils.serial import CLASS_INFO
+from trulens_eval.utils.pyschema import callable_name
+from trulens_eval.utils.pyschema import CLASS_INFO
 from trulens_eval.utils.serial import GetItemOrAttribute
 from trulens_eval.utils.serial import JSON
 from trulens_eval.utils.serial import JSON_BASES
 from trulens_eval.utils.serial import JSON_BASES_T
-from trulens_eval.utils.serial import json_str_of_obj
-from trulens_eval.utils.serial import jsonify
+from trulens_eval.utils.json import json_str_of_obj
+from trulens_eval.utils.json import jsonify
 from trulens_eval.utils.serial import JSONPath
-from trulens_eval.utils.serial import safe_signature
+from trulens_eval.utils.pyschema import safe_signature
 from trulens_eval.utils.serial import SerialModel
 from trulens_eval.utils.threading import TP
