+++ trulens_eval/trulens_eval/feedback/provider/endpoint/hugs.py
@@ -8,7 +8,7 @@ from trulens_eval.feedback.provider.endpoint.base import Endpoint
 from trulens_eval.feedback.provider.endpoint.base import EndpointCallback
 from trulens_eval.keys import _check_key
 from trulens_eval.keys import get_huggingface_headers
-from trulens_eval.util import WithClassInfo
+from trulens_eval.utils.pyschema import WithClassInfo
 
 
 class HuggingfaceCallback(EndpointCallback):
