+++ trulens_eval/trulens_eval/feedback/provider/hugs.py
@@ -7,7 +7,7 @@ import numpy as np
 from trulens_eval.feedback.provider.base import Provider
 from trulens_eval.feedback.provider.endpoint import HuggingfaceEndpoint
 from trulens_eval.feedback.provider.endpoint.base import Endpoint
-from trulens_eval.util import TP
+from trulens_eval.utils.threading import TP
 
 logger = logging.getLogger(__name__)
