+++ trulens_eval/trulens_eval/tru.py
@@ -18,8 +18,8 @@ from trulens_eval.feedback import Feedback
 from trulens_eval.schema import AppDefinition
 from trulens_eval.schema import FeedbackResult
 from trulens_eval.schema import Record
-from trulens_eval.util import SingletonPerName
-from trulens_eval.util import TP
+from trulens_eval.utils.python import SingletonPerName
+from trulens_eval.utils.threading import TP
 from trulens_eval.utils.notebook_utils import is_notebook
 from trulens_eval.utils.notebook_utils import setup_widget_stdout_stderr
 from trulens_eval.utils.text import UNICODE_CHECK
