+++ trulens_eval/trulens_eval/utils/langchain.py
@@ -10,13 +10,13 @@ from typing import List, Type
 
 from trulens_eval import app
 from trulens_eval.feedback import Feedback
-from trulens_eval.util import Class
-from trulens_eval.util import first
-from trulens_eval.util import JSON
-from trulens_eval.util import OptionalImports
-from trulens_eval.util import REQUIREMENT_LANGCHAIN
-from trulens_eval.util import second
-from trulens_eval.util import ThreadPoolExecutor
+from trulens_eval.utils.pyschema import Class
+from trulens_eval.utils.containers import first
+from trulens_eval.utils.serial import JSON
+from trulens_eval.utils.imports import OptionalImports
+from trulens_eval.utils.imports import REQUIREMENT_LANGCHAIN
+from trulens_eval.utils.containers import second
+from trulens_eval.utils.threading import ThreadPoolExecutor
 
 with OptionalImports(message=REQUIREMENT_LANGCHAIN):
     import langchain
