+++ trulens_eval/trulens_eval/utils/llama.py
@@ -11,13 +11,13 @@ from typing import List, Type
 from trulens_eval import app
 from trulens_eval import Feedback
 from trulens_eval.feedback import Feedback
-from trulens_eval.util import Class
-from trulens_eval.util import first
-from trulens_eval.util import JSON
-from trulens_eval.util import OptionalImports
-from trulens_eval.util import REQUIREMENT_LLAMA
-from trulens_eval.util import second
-from trulens_eval.util import TP
+from trulens_eval.utils.pyschema import Class
+from trulens_eval.utils.containers import first
+from trulens_eval.utils.serial import JSON
+from trulens_eval.utils.imports import OptionalImports
+from trulens_eval.utils.imports import REQUIREMENT_LLAMA
+from trulens_eval.utils.containers import second
+from trulens_eval.utils.threading import TP
 
 with OptionalImports(message=REQUIREMENT_LLAMA):
     from llama_index.indices.query.schema import QueryBundle
