+++ trulens_eval/trulens_eval/utils/trulens.py
@@ -6,8 +6,8 @@ Currently organizes all such components as "Other".
 from typing import Type
 
 from trulens_eval import app
-from trulens_eval.util import Class
-from trulens_eval.util import JSON
+from trulens_eval.utils.pyschema import Class
+from trulens_eval.utils.serial import JSON
 
 
 class Other(app.Other, app.TrulensComponent):
