+++ trulens_eval/trulens_eval/ux/components.py
@@ -13,12 +13,12 @@ from trulens_eval.schema import Metadata
 from trulens_eval.schema import Record
 from trulens_eval.schema import RecordAppCall
 from trulens_eval.schema import Select
-from trulens_eval.util import CLASS_INFO
-from trulens_eval.util import GetItemOrAttribute
-from trulens_eval.util import is_empty
-from trulens_eval.util import is_noserio
-from trulens_eval.util import jsonify
-from trulens_eval.util import JSONPath
+from trulens_eval.utils.pyschema import CLASS_INFO
+from trulens_eval.utils.serial import GetItemOrAttribute
+from trulens_eval.utils.containers import is_empty
+from trulens_eval.utils.pyschema import is_noserio
+from trulens_eval.utils.json import jsonify
+from trulens_eval.utils.serial import JSONPath
 
 
 def write_or_json(st, obj):
